static_cast directly from double to enum
Removed multistep casts that are no longer necessary. C++17 (N4659)
[8.2.9/10] says:

| A value of floating-point type can also be explicitly converted to an
| enumeration type. The resulting value is the same as converting the
| original value to the underlying type of the enumeration (7.10), and
| subsequently to the enumeration type.

+++ ihs_acctval.cpp
@@ -720,28 +720,28 @@ void AccountValue::SetInitialValues()
     Internal1035Amount = Outlay_->internal_1035_amount();
 
     ee_premium_allocation_method   = static_cast<oenum_allocation_method>
-        (static_cast<int>(Database_->Query(DB_EePremMethod))
+        (Database_->Query(DB_EePremMethod)
         );
     ee_premium_preferred_account   = static_cast<oenum_increment_account_preference>
-        (static_cast<int>(Database_->Query(DB_EePremAcct))
+        (Database_->Query(DB_EePremAcct)
         );
     er_premium_allocation_method   = static_cast<oenum_allocation_method>
-        (static_cast<int>(Database_->Query(DB_ErPremMethod))
+        (Database_->Query(DB_ErPremMethod)
         );
     er_premium_preferred_account   = static_cast<oenum_increment_account_preference>
-        (static_cast<int>(Database_->Query(DB_ErPremAcct))
+        (Database_->Query(DB_ErPremAcct)
         );
     deduction_method               = static_cast<oenum_increment_method>
-        (static_cast<int>(Database_->Query(DB_DeductionMethod))
+        (Database_->Query(DB_DeductionMethod)
         );
     deduction_preferred_account    = static_cast<oenum_increment_account_preference>
-        (static_cast<int>(Database_->Query(DB_DeductionAcct))
+        (Database_->Query(DB_DeductionAcct)
         );
     distribution_method            = static_cast<oenum_increment_method>
-        (static_cast<int>(Database_->Query(DB_DistributionMethod))
+        (Database_->Query(DB_DistributionMethod)
         );
     distribution_preferred_account = static_cast<oenum_increment_account_preference>
-        (static_cast<int>(Database_->Query(DB_DistributionAcct))
+        (Database_->Query(DB_DistributionAcct)
         );
 
     // If any account preference is the separate account, then a
